requirements.txt
openai>=1.50.0
playwright>=1.40.0
pillow>=10.0.0
python-dotenv>=1.0.0